LeagueSharp: Uploading docs from AttackableUnit to GameEndEventArgs.

+++ CollisionFlags.cs
@@ -2,13 +2,35 @@
 
 namespace LeagueSharp
 {
-  [Flags]
-  public enum CollisionFlags : ushort
-  {
-    None = (ushort) 0,
-    Grass = (ushort) 1,
-    Wall = (ushort) 2,
-    Building = (ushort) 64,
-    Prop = (ushort) 128,
-  }
+    /// <summary>
+    ///     Collision Flags, are used to define certain position in the game which hold a collision.
+    /// </summary>
+    [Flags]
+    public enum CollisionFlags : ushort
+    {
+        /// <summary>
+        ///     No Collision
+        /// </summary>
+        None = 0,
+
+        /// <summary>
+        ///     Grass Collision
+        /// </summary>
+        Grass = (ushort) 1,
+
+        /// <summary>
+        ///     Wall Collision
+        /// </summary>
+        Wall = (ushort) 2,
+
+        /// <summary>
+        ///     Building Collision
+        /// </summary>
+        Building = (ushort) 64,
+
+        /// <summary>
+        ///     Prop (Object) Collision
+        /// </summary>
+        Prop = (ushort) 128
+    }
 }
